// utils/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json }
  | Json[]

export interface Database {
  public: {
    Tables: {
      decks: {
        Row: {
          deck_id: string
          deck_image_url: string | null
          deck_name: string | null
          elo: number | null
          tier: string | null
          time_added: string
        }
        Insert: {
          deck_id?: string
          deck_image_url?: string | null
          deck_name?: string | null
          elo?: number | null
          tier?: string | null
          time_added?: string
        }
        Update: {
          deck_id?: string
          deck_image_url?: string | null
          deck_name?: string | null
          elo?: number | null
          tier?: string | null
          time_added?: string
        }
      }
      profiles: {
        Row: {
          avatar_url: string | null
          full_name: string | null
          id: string
          updated_at: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          full_name?: string | null
          id: string
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          full_name?: string | null
          id?: string
          updated_at?: string | null
          username?: string | null
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      delete_avatar: {
        Args: {
          avatar_url: string
        }
        Returns: Record<string, unknown>
      }
      delete_storage_object: {
        Args: {
          bucket: string
          object: string
        }
        Returns: Record<string, unknown>
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}